requirements.txt
flask
gunicorn
opencv-python-headless
numpy
pillow
scikit-learn
tensorflow-cpu
requests
qrcode
pyzbar
tldextract
google-genai
python-dotenv